ref(nav) break down primary nav
